Enamul

See merge request iam00/way_housing_crm!39

+++ resources/views/negotiation_analysis/negotiation_analysis_list.blade.php
@@ -9,11 +9,19 @@
                 <div class="col-12">
                     <div class="page-title-box d-flex align-items-center justify-content-between">
                         <h4 class="mb-sm-0">Negotiations</h4> 
+                        <p class="d-none">Employee: {{auth()->user()->name}}</p> 
+                        <input type="hidden" id="hideExport" value=":nth-child(1),:nth-child(2)"> 
+                        <input type="hidden" id="pageSize" value="legal">
+                        <input type="hidden" id="fontSize" value="8"> 
                         <div class="page-title-right">
-                            <ol class="breadcrumb m-0">
-                                <li class="breadcrumb-item"><a href="javascript: void(0);">Dashboard</a></li>
-                                <li class="breadcrumb-item active">Negotiations</li>
-                            </ol>
+                            <div class="dt-buttons flex-wrap mb-2">     
+                                <button class="btn btn-secondary mr-3" data-bs-toggle="modal" data-bs-target="#wayting_day_setting">
+                                    <span><i class="fas fa-cog"></i> Waiting Day</span>
+                                </button>  
+                                <button class="btn btn-secondary" data-bs-toggle="offcanvas" data-bs-target="#offcanvas">
+                                    <span><i class="fas fa-filter"></i> Filter</span>
+                                </button> 
+                            </div>
                         </div>
 
                     </div>
@@ -24,33 +32,8 @@
             <div class="row">
                 <div class="col-12">
                     <div class="card"> 
-                        <div class="card-body">
-
-                            <div class="d-flex justify-content-between"> 
-                                <div class="">
-                                    <div class="dt-buttons btn-group flex-wrap mb-2">      
-                                        <button class="btn btn-primary buttons-copy buttons-html5" tabindex="0" aria-controls="datatable-buttons" type="button">
-                                            <span><i class="fas fa-file-excel"></i> Excel</span>
-                                        </button>
-
-                                        <button class="btn btn-secondary buttons-excel buttons-html5" tabindex="0" aria-controls="datatable-buttons" type="button">
-                                            <span><i class="fas fa-file-pdf"></i> PDF</span>
-                                        </button> 
-                                    </div> 
-                                </div>
-                                <div class="">
-                                    <div class="dt-buttons flex-wrap mb-2">     
-                                        <button class="btn btn-secondary mr-3" data-bs-toggle="modal" data-bs-target="#wayting_day_setting">
-                                            <span><i class="fas fa-cog"></i> Waiting Day</span>
-                                        </button>  
-                                        <button class="btn btn-secondary" data-bs-toggle="offcanvas" data-bs-target="#offcanvas">
-                                            <span><i class="fas fa-filter"></i> Filter</span>
-                                        </button> 
-                                    </div>
-                                </div>
-                           </div>
-
-                            <table class="table table-hover table-bordered table-striped dt-responsive nowrap fs-10" style="border-collapse: collapse; border-spacing: 0; width: 100%;">
+                        <div class="card-body">  
+                            <table id="datatable" class="table table-hover table-bordered table-striped dt-responsive nowrap fs-10" style="border-collapse: collapse; border-spacing: 0; width: 100%;">
                                 <thead>
                                     <tr>
                                         <th>Action</th>
@@ -67,28 +50,41 @@
                                 </thead>
                                 <tbody> 
                                     @foreach ($negotiations as  $negotiation)
-                                    <tr>
+                                    <tr class="{{$negotiation->approve_by==null?"table-warning":""}}">
                                         <td class="text-center" data-bs-toggle="tooltip" title="Action"> 
                                             <div class="dropdown">
-                                                <a href="javascript:void(0)" data-bs-toggle="dropdown" aria-expanded="false"><i class="fas fa-ellipsis-v align-middle ms-2 cursor-pointer"></i></a>
+                                                <a href="javascript:void(0)" data-bs-toggle="dropdown" aria-expanded="false">
+                                                    <img class="rounded avatar-2xs p-0" src="{{@$negotiation->customer->user->image()}}">
+                                                </a>
                                                 <div class="dropdown-menu dropdown-menu-animated">
-                                                    <a class="dropdown-item" href="customer_profile.html">Customer Profile</a> 
-                                                    <a class="dropdown-item" href="{{route('negotiation-analysis.edit',$negotiation->id)}}">Edit</a>
-                                                    <a class="dropdown-item" href="javascript:void(0)" onclick="deleteItem('{{ route('negotiation.delete',$negotiation->id) }}')">Delete</a> 
-                                                    <a class="dropdown-item" href="{{route('salse.create',['customer'=>$negotiation->customer->id])}}">Sales Create</a>
+                                                    <a class="dropdown-item" href="{{route('customer.profile',$negotiation->customer_id)}}">Customer Profile</a>
+                                                    @if ($negotiation->approve_by==null)
+                                                        @can('negotiation-analysis-manage')
+                                                            <a class="dropdown-item" href="{{route('negotiation-analysis.edit',$negotiation->id)}}">Edit</a>
+                                                        @endcan 
+                                                    @endif  
+
+                                                    @can('negotiation-analysis-delete')
+                                                        <a class="dropdown-item" href="javascript:void(0)" onclick="deleteItem('{{ route('negotiation.delete',$negotiation->id) }}')">Delete</a> 
+                                                    @endcan
+                                                    
+                                                    @if ($negotiation->approve_by != null)
+                                                        @can('sales-manage')
+                                                            <a class="dropdown-item" href="{{route('salse.create',['customer'=>$negotiation->customer->id])}}">Sales Create</a>
+                                                        @endcan 
+                                                    @endif
                                                 </div>
                                             </div> 
                                         </td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}">{{ $loop->iteration}}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}">{{ $negotiation->created_at }}</td>
-                                       
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}">{{ @$negotiation->customer->user->name }}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}"> {{ @$negotiation->customer->user->phone }}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}"> {{ @$negotiation->customer->user->userAddress->address }}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}"> {{ @$negotiation->negotiation_amount }}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}"> {{ @$negotiation->project->name }}</td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}">  2 #dummmy </td>
-                                        <td class="{{ $negotiation->status == 0 ? 'text-danger' : '' }}">  {{ @$negotiation->employee->user->name }} </td>
+                                        <td class="">{{ $loop->iteration}}</td>
+                                        <td class="">{{ get_date($negotiation->created_at) }}</td> 
+                                        <td class="">{{ @$negotiation->customer->user->name }}</td>
+                                        <td class=""> {{ @$negotiation->customer->user->phone }}</td>
+                                        <td class=""> {{ @$negotiation->customer->user->userAddress->address }}</td>
+                                        <td class=""> {{ get_price(@$negotiation->negotiation_amount) }}</td>
+                                        <td class=""> {{ @$negotiation->project->name }}</td>
+                                        <td class="">{{count(json_decode($negotiation->project_units))}} </td>
+                                        <td class="">{{ @$negotiation->customer->reference->name }} [{{ @$negotiation->customer->reference->user_id }}] </td>
                                     </tr> 
                                 @endforeach
                                 </tbody>
@@ -271,17 +267,18 @@
             <div class="modal-header"> 
                 <h5 class="modal-title">Update Setting</h5><button type="button" class="btn btn-sm btn-label-danger btn-icon" data-bs-dismiss="modal"><i class="mdi mdi-close"></i></button>
             </div>
-            <form action=""> 
+            <form action="{{route('update.negotiation.waiting.day')}}" method="post">
+                @csrf 
                 <div class="modal-body"> 
                     <div class="form-group mb-2">
                         <label for="waiting_day">Waiting Day<span class="text-danger">*</span></label>
-                        <input type="number" class="form-control" id="waiting_day" name="waiting_day" value="7" required>
+                        <input type="number" class="form-control" id="waiting_day" min="1" name="waiting_day" value="{{@$waiting_day->waiting_day}}" required>
                     </div>   
                 </div>
 
                 <div class="modal-footer">
                     <div class="text-end">
-                        <button class="btn btn-primary"><i class="fas fa-save"></i> Submit</button> <button class="btn btn-outline-danger"><i class="mdi mdi-refresh"></i> Reset</button>
+                        <button type="submit" class="btn btn-primary"><i class="fas fa-save"></i> Submit</button> <button type="button" class="btn btn-outline-danger refresh_btn"><i class="mdi mdi-refresh"></i> Reset</button>
                     </div> 
                 </div> 
             </form>
@@ -291,6 +288,7 @@
 @endsection 
 
 @section('script')
+@include('includes.data_table')
 <script>
     getDateRange('join_date');
     getDateRange('last_cold_calling');
